requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
scikit-learn
joblib
requests